netContain: check working directory & cd

+++ scripts/netContain/build_image.sh
@@ -38,6 +38,14 @@ function build_image_from_path() {
 set -x
 set -e
 
+# ensure this script wasn't called from the directory where this script
+# lives; it should be called from the repository's top level
+script_dir="$(dirname -- "$0")"
+if [ "$script_dir" == "." ]; then
+	echo "This script must be called from the top level of the repository"
+	exit 1
+fi
+
 rm -rf scripts/netContain/bin
 mkdir scripts/netContain/bin
